Move telemetry to common/vscode
