+++ extensions/ql-vscode/src/compare/compare-view.ts
@@ -23,7 +23,7 @@ import {
   AbstractWebview,
   WebviewPanelConfig,
 } from "../common/vscode/abstract-webview";
-import { telemetryListener } from "../telemetry";
+import { telemetryListener } from "../common/vscode/telemetry";
 import { redactableError } from "../pure/errors";
 
 interface ComparePair {
